% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citest_utils.R
\name{extractValidCITestResults}
\alias{extractValidCITestResults}
\title{TODO: make citestResults an object with both the labels and the data.frame}
\usage{
extractValidCITestResults(citestResults, cur_varnames, new_varnames)
}
\description{
TODO: make citestResults an object with both the labels and the data.frame
}
