% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isMConTriplet}
\alias{isMConTriplet}
\title{if definite=TRUE, it only returns FALSE if the triplet is of definite status
and it is either a collider such that no descendants of it in in Z, or a
definite non-collider.}
\usage{
isMConTriplet(
  amat,
  vi_name,
  vm_name,
  vj_name,
  Z,
  definite = TRUE,
  verbose = FALSE
)
}
\description{
if definite=TRUE, it only returns FALSE if the triplet is of definite status
and it is either a collider such that no descendants of it in in Z, or a
definite non-collider.
}
